Rename project (#17)
* Rename CvBuilder to Kamisaku and update all references

Rebranded the project by renaming the main module, file structure, and related references from CvBuilder to Kamisaku. Adjusted gemspec, test files, documentation, and resource names to reflect the new project name.

* Update version and fix changelog URI in gemspec

Bump the version from 0.1.0 to 0.1.1 to prepare for a new release. Correct the changelog_uri in the gemspec to point to the accurate file location on GitHub.

* Update kamisaku gem to version 0.1.1

* Add Gem Version badge to README